It works.
it takes in a folder and a destanation and then copys and pastes every 5 min. I have to make it save the values permantly so that you don't have to enter them again on every reboot and I will make a linux one that has / instead of \ or maybe I will have it do it dynamicly. That would be cool.