Fixed a variety of problems
Now it does not have the isstue where it will throw a permission error and it also now handles other errors properly. I plan to put logging in for better future diagnosis in the future along with working to make it so that you can use absoloute paths in the src folder. Right now you have to use a reletave path requireing you to have the file in for example your User's folder. That seems to be the best place to keep it at this time.